[Bugfix][CI/Build] Fix test and improve code for `merge_async_iterators` (#5096)
